Single listen thread proof of concept.
